Server class reorganization
