+++ src/socket/Server.cpp
@@ -6,12 +6,10 @@
 /*   By: mamartin <mamartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/08 18:07:44 by mamartin          #+#    #+#             */
-/*   Updated: 2021/10/18 02:29:38 by mamartin         ###   ########.fr       */
+/*   Updated: 2021/10/18 04:08:49 by mamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <sstream>
-#include <cerrno>
 #include "Server.hpp"
 
 Server::Server(const Config& conf) :
@@ -125,29 +123,14 @@ Server::handle_request(Client& client)
 	if (_handle_redirection(client, route) == true)
 		return (OK); // redirection response has been created, ready to send
 
-	// delete specified file
-	if (req.getMethod() == "DELETE")
-	{
-		_delete_file(client, route, uri.getPath());
-		return (OK);
-	}
-
-	if (_check_cgi_extension(route, uri.getPath()))
-	{
-		// _handle_cgi(uri, client);
-	}
-
-	if (_handle_upload(client, route) == true) // file has been created or an error occured
-		return (OK); // client state == (IDLE || WAITING_ANSWER)
-	
-	// search content requested by the client
-	code = _find_resource(route, uri.getPath(), client);
-	if (code != OK)
-		return (_handle_error(client, code));
-	else if (client.file()) // resource found
-		client.state(IDLE);
-
-	return (OK);
+	// call function associated to the request method
+	char method_identifier = req.getMethod()[0];
+	if (method_identifier == 'G') // GET
+		return (_handle_get(client, route, uri));
+	else if (method_identifier == 'P') // POST
+		return (_handle_post(client, route, uri));
+	else // DELETE
+		return (_handle_delete(client, route, uri));
 }
 
 void
@@ -159,7 +142,7 @@ Server::send_response(Client& client)
 	{
 		client.write_trials++;
 		std::cerr << "server > sending client response failed: " << strerror(errno) << "\n";
-		if (client.write_trials == 5)
+		if (client.write_trials == 5) // too much fails
 		{
 			std::cerr << "server > client disconnected.\n";
 			client.state(DISCONNECTED);
@@ -175,7 +158,7 @@ Server::send_response(Client& client)
 
 	try
 	{
-		client.clear();
+		client.clear(); // clear request, response, state...
 	}
 	catch (const std::exception& e)
 	{
@@ -247,475 +230,4 @@ const Listener&
 Server::get_listener(void) const
 {
 	return (_host);
-}
-
-/*** PRIVATE FUNCTIONS ********************************************************/
-
-bool
-Server::_read_request(Client &client)
-{
-	char	buffer[BUFFER_SIZE];
-	int		readBytes	= read(client.fd(), &buffer, BUFFER_SIZE);
-
-	if (readBytes < 0)
-	{
-		std::cerr << "server > reading client request failed: " << strerror(errno) << "\n";
-		_handle_error(client, INTERNAL_SERVER_ERROR);
-		return (false);
-	}
-	else if (readBytes == 0)
-		client.state(DISCONNECTED);
-	else
-	{
-		buffer[readBytes] = '\0';
-		client.request().parseChunk(std::string(buffer));
-	}
-	return (client.request().isReady());
-}
-
-const Route&
-Server::_resolve_routes(const std::string& uri_path)
-{
-	const Route* matching = NULL;
-
-	for (std::vector<Route>::const_iterator it = _config.routes.begin();
-			it != _config.routes.end();
-			++it)
-	{
-		if (uri_path.find(it->location) == 0) // prefix found in uri
-		{
-			if (!matching)
-				matching = &(*it); // first prefix matching
-			else if (matching->location.length() < it->location.length())
-				matching = &(*it); // longest prefix to match
-		}
-	}
-	return (*matching);
-}
-
-int
-Server::_check_request_validity(const Route& rules, HTTPRequest& request)
-{
-	// check HTTP protocol version
-	if (request.getVersion() != "HTTP/1.1")
-		return (HTTP_VERSION_NOT_SUPPORTED);
-
-	// check host header field
-	if (!request.getHost().length())
-		return (BAD_REQUEST);
-
-	// check that the specified method is implemented by our server
-	bool method = false;
-	for (std::set<std::string>::iterator it = rules._methods_supported.begin();
-		it != rules._methods_supported.end();
-		++it)
-	{
-		if (*it == request.getMethod())
-		{
-			method = true;
-			break ;
-		}
-	}
-	if (!method)
-		return (NOT_IMPLEMENTED);
-
-	// check that the specified method is allowed by the route rules
-	method = false;
-	for (std::set<std::string>::const_iterator it = rules.methods.begin();
-			it != rules.methods.end();
-			++it)
-	{
-		if (*it == request.getMethod())
-		{
-			method = true;
-			break ;
-		}
-	}
-	if (!method)
-		return (METHOD_NOT_ALLOWED);
-
-	// a body limit of 0 means unlimited
-	if (_config.body_limit && request.getBodySize() > _config.body_limit)
-		return (PAYLOAD_TOO_LARGE);
-
-	// POST requests MUST have a Content-Length header
-	if (request.getMethod() == "POST" && !request.getHeader().getValue("Content-Length"))
-		return (LENGTH_REQUIRED);
-	return (OK);
-}
-
-int
-Server::_check_cgi_extension(const Route& rules, const std::string& uri_path)
-{
-	if (rules.cgi_extension.length() == 0)
-		return (false); // no cgi
-
-	// look for cgi_extension in uri path
-	size_t	pos = uri_path.rfind(rules.cgi_extension);
-	if (pos == std::string::npos)
-		return (false);
-	else if (pos == uri_path.length() - rules.cgi_extension.length())
-		return (true); // cgi_extension found at the end of path
-	return (false); // not found or not where it should be
-}
-
-int
-Server::_find_resource(const Route& rules, std::string path, Client& client)
-{
-	// build path from root dir and uri path
-	path.erase(0, rules.location.length());		// location/path/to/file -> /path/to/file
-	path = _append_paths(rules._root, path);	// root/path/to/file
-
-	struct stat	pathinfo;
-	if (stat(path.data(), &pathinfo) == -1)
-	{
-		if (errno == ENOENT || errno == ENOTDIR)
-			return (NOT_FOUND); // path doesn't exist
-		else
-		{
-			std::cerr << "server > stat() failed: " << strerror(errno) << "\n";
-			return (INTERNAL_SERVER_ERROR); // other error
-		}
-	}
-	
-	// path is a directory
-	if (pathinfo.st_mode & S_IFDIR)
-	{
-		DIR*						dirptr;
-		struct dirent*				file;
-		std::vector<struct dirent>	dirls;
-
-		// open directory
-		if (!(dirptr = opendir(path.data())))
-		{
-			std::cerr << "server > cannot open directory \"" << path << "\": " << strerror(errno) << "\n";
-			return (INTERNAL_SERVER_ERROR);
-		}
-
-		errno = 0;
-		while ((file = readdir(dirptr))) // read directory entries
-		{
-			if (_is_index_file(rules, file))
-				break ; // index file found !
-			dirls.push_back(*file);
-		}
-		closedir(dirptr);
-		if (errno) // readdir failed
-		{
-			std::cerr << "server > cannot read directory \"" << path << "\": " << strerror(errno) << "\n";
-			return (INTERNAL_SERVER_ERROR);
-		}
-
-		if (!file) // index file not found...
-		{
-			if (rules._autoindex) // autoindex enabled
-			{
-				client.response().gen_autoindex(dirls, path);
-				_create_response(client);
-				return (OK);
-			}
-			return (NOT_FOUND);
-		}
-		else
-			path = _append_paths(path, file->d_name); // path + filename
-	}
-	
-	if (_file_already_requested(client, path))
-		return (OK); // another client asked for the same file
-
-	// open file
-	int fd = open(path.data(), O_RDONLY | O_NONBLOCK);
-	if (fd == -1)
-	{
-		std::cerr << "server > cannot open file \"" << path << "\": " << strerror(errno) << "\n";
-		return (INTERNAL_SERVER_ERROR);
-	}
-
-	_files.push_back(f_pollfd(path, fd));
-	client.file(&_files.back());
-	return (OK);
-}
-
-bool
-Server::_is_index_file(const Route& rules, struct dirent* file)
-{
-	for (std::set<std::string>::const_iterator it = rules._index_paths.begin();
-			it != rules._index_paths.end();
-			++it)
-	{
-		if (file->d_type == DT_REG) // regular file
-		{
-			if (*it == file->d_name)
-				return (true); // file is an index file
-		}
-	}
-	return (false); // not an index file
-}
-
-int
-Server::_handle_error(Client& client, int status, bool autogen)
-{
-	if (!autogen) // search error page in server configuration
-	{
-		std::map<int, std::string>::const_iterator	errpage;
-	
-		errpage = _config._error_pages.find(status);
-		if (errpage != _config._error_pages.end()) // page exists
-		{
-			client.response().setStatus((status_code)status);
-			_create_response(client, &errpage->second);
-			return (0);
-		}
-		else // doesn't exist
-			return (_handle_error(client, status, true)); // retry with auto-generation enabled
-	}
-	else // generate page
-	{
-		client.response().gen_error_page(status);
-		_create_response(client);
-	}
-	return (0);
-}
-
-bool
-Server::_file_already_requested(Client& client, std::string const &filepath)
-{
-	for (size_t i = 0; i < _files.size(); i++)
-	{
-		if (_files[i].name == filepath) // file already in server's list
-		{
-			// assign file to the client
-			client.file(&_files[i]);
-			return (true);
-		}
-	}
-	return (false);
-}
-
-std::string
-Server::_append_paths(const std::string& str1, const std::string& str2)
-{
-	size_t		last_index = str1.length() - 1;
-	std::string	ret;
-
-	if (str1[last_index] == '/' || str2[0] == '/')
-		ret = str1 + str2;
-	else
-		ret = str1 + "/" + str2;
-	return (ret);
-}
-
-void
-Server::_create_response(Client& client, const std::string *body)
-{
-	/*
-		no need to set response status here because
-			it's either 200 (default value)
-			or already set by Server::_handle_error()
-
-		Body is either sent by create_file_response() -> body pointer
-			or generated by Server::_handle_error() and Server::_find_resource()
-
-		If you want more headers I can add them
-			retry-after and last-modified are some examples
-		Content-Type is always text/html
-			it will be changed when configuration will includes
-			a map of MIME types
-		Connection is keep-alive by default
-	*/ 
-
-	HTTPResponse&		response = client.response();
-	HTTPHeader			headers;
-	std::stringstream	ss;
-	
-	if (body)
-		response.setBody(*body);
-	ss << response.getBodySize();
-
-	// set headers
-	response.setContentType("text/html");					// default Content-Type
-	headers.addValue("Content-Length", ss.str());			// Content-Length
-	if (response.getStatus() == BAD_REQUEST)
-		headers.addValue("Connection", "close");			// Connection (nginx behavior)
-	else if (response.getStatus() == METHOD_NOT_ALLOWED)	// Allow (only for 405 errors)
-	{
-		std::string	allow_header_val;
-
-		for (std::set<std::string>::iterator it = client.rules()->methods.begin();
-			it != client.rules()->methods.end();
-			++it)
-				allow_header_val += *it;
-		headers.addValue("Allow", allow_header_val);
-	}
-	else if (response.getStatus() / 100 == 3) // status indicates a redirection
-	{
-		const std::string&	new_url = client.rules()->redirection.second;
-
-		// add a Location header with the new url
-		// url may contain variables like $host for exemple
-		// we need to replace them by their actual values
-		headers.addValue("Location", _replace_conf_vars(client, new_url));
-	}
-	else if (response.getStatus() == CREATED)
-	{
-		std::string ref = _get_uri_reference(client.file()->name);
-		if (ref.length()) // new file can be referenced as an uri
-			headers.addValue("Location", ref); // add this uri reference to the response
-	}
-	response.setHeader(headers);
-
-	response.setReady(true);
-	client.rules(NULL);
-	client.state(WAITING_ANSWER);
-}
-
-bool
-Server::_handle_redirection(Client& client, const Route& rules)
-{
-	if (!rules.redirection.first)
-		return (false); // no redirection configured
-
-	client.response().setStatus((status_code)rules.redirection.first);
-	_create_response(client);
-	return (true);
-}
-
-std::string
-Server::_replace_conf_vars(Client& client, const std::string& redirection)
-{
-	std::string	url = redirection;
-	std::string	port;
-	size_t		pos;
-
- 	// replace "$host" placeholder by real hostname
-	pos = url.find("$host");
-	if (pos != std::string::npos)
-	{
-		std::string	host		= client.request().getHost();
-		size_t		port_pos	= host.rfind(":");
-
-		// remove port segment if any
-		if (port_pos != std::string::npos) 
-			host.erase(port_pos, host.length() - port_pos);
-		url.replace(pos, 5, host);
-	}
-
- 	// replace "$server_port"
-	pos = url.find("$server_port");
-	if (pos != std::string::npos)
-	{
-		std::stringstream	ss;
-
-		ss << ntohs(_config.port);
-		url.replace(pos, 12, ss.str());
-	}
-
- 	// replace "$uri"
-	pos = url.find("$uri");
-	if (pos != std::string::npos)
-		url.replace(pos, 4, client.request().getURI().getPath());
-	return (url);
-}
-
-bool
-Server::_handle_upload(Client& client, const Route& rules)
-{
-	if (client.request().getMethod() != "POST")
-		return (false); // uploads use POST method only
-
-	(void)rules;
-
-	const std::string*	content_type	= client.request().getContentType();
-	std::string			mime_type		= "application/octet-stream"; // default content-type if not provided
-	std::string			boundary		= "";
-
-	if (*content_type == "multipart/form-data")
-	{
-		mime_type = *content_type;
-		return (false);
-	}
-	else if (*content_type == "application/x-www-form-urlencoded")
-		return (false); // this content-type cannot be used for file transfer
-	else
-	{
-		std::string filename = client.request().getURI().getPath();
-
-		// build file path
-		filename.erase(0, rules.location.length());				// remove location prefix
-		filename = _append_paths(rules.upload_path, filename);	// filename = upload_path + filename
-		
-		// create file
-		client.file(_create_file(filename, client.request().getBody()));
-		if (!client.file()) // an error occured when trying to open file (see error logs for more details)
-		{
-			_handle_error(client, INTERNAL_SERVER_ERROR);
-			return (true);
-		}
-	}
-
-	client.state(IDLE);
-	return (true);
-}
-
-f_pollfd*
-Server::_create_file(const std::string& filename, const std::string& data)
-{
-	int	mode = S_IRWXU | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH; // 755 filemode
-	int	fd;
-
-	// open file (read only, non blocking fd, create it if it doesn't already exist)
-	fd = open(filename.data(), O_WRONLY | O_CREAT | O_NONBLOCK | O_APPEND, mode);
-	if (fd == -1)
-	{
-		std::cerr << "server > cannot open file \"" << filename << "\": " << strerror(errno) << "\n";
-		return (NULL);
-	}
-
-	_files.push_back(f_pollfd(filename, fd, POLLOUT, data));
-	return (&_files.back());
-}
-
-std::string
-Server::_get_uri_reference(const std::string& filename)
-{
-	std::string	ref = filename;
-
-	for (std::vector<Route>::const_iterator it = _config.routes.begin();
-			it != _config.routes.end();
-			++it)
-	{
-		if (filename.find(it->_root) == 0) // root path found in filename
-		{
-			// replace root by location in filename
-			ref.erase(0, it->_root.length());
-			ref = _append_paths(it->location, ref);
-			return (ref); // return uri reference to filename
-		}
-	}
-	return (""); // filename cannot be referenced as an uri
-}
-
-void
-Server::_delete_file(Client& client, const Route& rules, const std::string& uri_path)
-{
-	std::string path = uri_path;
-
-	// build path from root dir and uri path
-	path.erase(0, rules.location.length());		// location/path/to/file -> /path/to/file
-	path = _append_paths(rules._root, path);	// root/path/to/file
-
-	if (remove(path.data()) == -1)
-	{
-		std::cerr << "server > deletion of \"" << path << "\" failed: " << strerror(errno) << "\n";
-		if (errno == EACCES || errno == EROFS || errno == EPERM)
-			_handle_error(client, FORBIDDEN);
-		else if (errno == ENOENT || errno == ENOTDIR)
-			_handle_error(client, NOT_FOUND);
-		else
-			_handle_error(client, INTERNAL_SERVER_ERROR);
-		return ;
-	}
-		
-	client.response().setStatus(NO_CONTENT);
-	_create_response(client);
 }
